notification tick listener

+++ Assets/Sources/Features/Time/TickComponent.cs
@@ -3,5 +3,5 @@ using Entitas.CodeGenerator;
 
 [Core, SingleEntity]
 public class TickComponent : IComponent {
-	public float currentTick;	
+	public long currentTick;
 }
